refactor(intelligence,http): move semantic_drift handler to mockforge-intelligence (#555 phase 3)
The semantic-drift HTTP surface depends on three intelligence-side
modules and one foundation-side type module:

  - ai_contract_diff::{ContractDiffAnalyzer, ContractDiffConfig}
    (moved to intelligence in #562 phase 4)
  - incidents::semantic_manager::{SemanticIncident, SemanticIncidentManager}
    (moved to intelligence in #562 phase 9)
  - database::Database (moved to intelligence in #611)
  - mockforge_foundation::incidents_types::{IncidentSeverity, IncidentStatus}

With the #611 Database prereq in, the rewrite is mechanical:
`mockforge_core::*` → `crate::*` for the two intelligence-side paths,
`mockforge_core::incidents::types::*` →
`mockforge_foundation::incidents_types::*` for the inline types
references in `map_row_to_semantic_incident` and the status parser.

## What changed

- `git mv` http/src/handlers/semantic_drift.rs ->
  intelligence/src/handlers/semantic_drift.rs.
- 5 in-file path rewrites:
    - mockforge_core::ai_contract_diff::* → crate::ai_contract_diff::*
    - mockforge_core::incidents::semantic_manager::* →
      crate::incidents::semantic_manager::*
    - mockforge_core::incidents::types::IncidentSeverity →
      mockforge_foundation::incidents_types::IncidentSeverity
    - mockforge_core::incidents::types::IncidentStatus →
      mockforge_foundation::incidents_types::IncidentStatus
      (used in the `match status_str.as_str()` block)
- intelligence/src/handlers/mod.rs: + `pub mod semantic_drift;` with
  a docstring noting which prior #562 phases unblocked it.
- http/src/handlers/semantic_drift.rs is now a 1-line re-export shim
  so router code in `mockforge_http::lib` keeps resolving.

## Drive-by: drop the leftover `#[cfg(feature = "database")]` gate on
`pub mod database;` in intelligence's lib.rs

#611 landed with `#[cfg(feature = "database")] pub mod database;` —
but the inline re-export shim in `mockforge_http::lib` references
`mockforge_intelligence::database` unconditionally. Without the
`database` feature on http, the shim fails to resolve and downstream
integration tests stop compiling. The `Database` type itself already
compiles in both feature states (the sqlx-bearing pool field is
`#[cfg]`-gated internally, matching the original
`mockforge_http::database` shape), so the module-level gate is the
wrong layer. Removing it here.

## Verification

- cargo check -p mockforge-intelligence -p mockforge-http — clean (without feature)
- cargo check -p mockforge-intelligence -p mockforge-http --features mockforge-http/database — clean
- cargo test -p mockforge-intelligence -p mockforge-http --lib --features mockforge-http/database — 926 pass
- cargo clippy ... --features mockforge-http/database -- -D warnings — clean
- cargo fmt --all --check — clean
